Include world map

+++ src/constants.hpp
@@ -1,4 +1,4 @@
 #pragma once
-constexpr double PI = 3.14159;
+constexpr float PI = 3.14159f;
 constexpr float RAD = PI/180;
 constexpr const char* RES_PATH = "resources/";
